requirements.txt
# 104爬蟲 - 本機執行版本所需套件

# 核心套件
scrapy>=2.11.0

# 環境變數管理
python-dotenv>=1.0.0

# 資料處理 (選用,如果需要進階資料處理)
pandas>=2.0.0

# ==========================================
# 以下是進階功能套件(預設不需要安裝)
# ==========================================

# 如果你想用Elasticsearch儲存:
# elasticsearch>=8.9.0
# elasticsearch-dsl>=8.9.0

# 如果你想用MongoDB儲存:
# pymongo>=4.5.0
